Race condition and clearAllMocks check. Timestamp=1771546672
